Добавление подтверждения удаления задачи views.py и создание delete_confirmation.html

+++ source/webapp/views.py
@@ -57,6 +57,11 @@ def update_task_view(request,pk):
 
 
 def delete(request,pk):
-    task = Task.objects.get(pk=pk)
-    task.delete()
+    task = get_object_or_404(Task, pk=pk)
+    if request.method == 'GET':
+        form = TaskForm(data={'description': task.description, 'status': task.status,
+                              'text': task.text, 'completed_at': task.completed_at})
+        return render(request, 'delete_confirmation.html', context={'form':form,'task':task})
+    elif request.method == 'POST':
+        task.delete()
     return redirect('index')
